+++ src/sqlfluff/templaters.py
@@ -1,5 +1,6 @@
 """Defines the templaters."""
 
+import os.path
 import ast
 
 from .errors import SQLTemplaterError
@@ -158,7 +159,7 @@ class JinjaTemplateInterface(PythonTemplateInterface):
     name = 'jinja'
 
     @staticmethod
-    def _extract_macros_from_template(template, env):
+    def _extract_macros_from_template(template, env, ctx):
         """Take a template string and extract any macros from it.
 
         Lovingly inspired by http://codyaray.com/2015/05/auto-load-jinja2-macros
@@ -167,7 +168,7 @@ class JinjaTemplateInterface(PythonTemplateInterface):
 
         # Iterate through keys exported from the loaded template string
         context = {}
-        macro_template = env.from_string(template)
+        macro_template = env.from_string(template, globals=ctx)
         # This is kind of low level and hacky but it works
         for k in macro_template.module.__dict__:
             attr = getattr(macro_template.module, k)
@@ -177,7 +178,36 @@ class JinjaTemplateInterface(PythonTemplateInterface):
         # Return the context
         return context
 
-    def _extract_macros_from_config(self, config, env):
+    @classmethod
+    def _extract_macros_from_path(cls, path, env, ctx):
+        """Take a path and extract macros from it."""
+        # Does the path exist? It should as this check was done on config load.
+        if not os.path.exists(path):
+            raise ValueError("Path does not exist: {0}".format(path))
+
+        macro_ctx = {}
+        if os.path.isfile(path):
+            # It's a file. Extract macros from it.
+            with open(path, 'r') as opened_file:
+                template = opened_file.read()
+            # Update the context with macros from the file.
+            macro_ctx.update(
+                cls._extract_macros_from_template(
+                    template, env=env, ctx=ctx
+                )
+            )
+        else:
+            # It's a directory. Iterate through files in it and extract from them.
+            for dirpath, _, files in os.walk(path):
+                for fname in files:
+                    if fname.endswith('.sql'):
+                        macro_ctx.update(cls._extract_macros_from_path(
+                            os.path.join(dirpath, fname),
+                            env=env, ctx=ctx
+                        ))
+        return macro_ctx
+
+    def _extract_macros_from_config(self, config, env, ctx):
         """Take a config and load any macros from it."""
         if config:
             # This is now a nested section
@@ -190,7 +220,7 @@ class JinjaTemplateInterface(PythonTemplateInterface):
         for value in loaded_context.values():
             macro_ctx.update(
                 self._extract_macros_from_template(
-                    value, env=env
+                    value, env=env, ctx=ctx
                 )
             )
         return macro_ctx
@@ -217,13 +247,42 @@ class JinjaTemplateInterface(PythonTemplateInterface):
             autoescape=False, extensions=['jinja2.ext.do']
         )
 
-        ctx = self._extract_macros_from_config(config=config, env=env)
-        # Apply to globals
-        env.globals.update(ctx)
+        if not config:
+            raise ValueError("For the jinja templater, the `process()` method requires a config object.")
 
-        template = env.from_string(in_str)
+        # Load the context
         live_context = self.get_context(fname=fname, config=config)
+        # Apply dbt builtin functions if we're allowed.
+        apply_dbt_builtins = config.get_section((self.templater_selector, self.name, 'apply_dbt_builtins'))
+        if apply_dbt_builtins:
+            # This feels a bit wrong defining these here, they should probably
+            # be configurable somewhere sensible. But for now they're not.
+            # TODO: Come up with a better solution.
+            this_fname = fname or 'this_model'
+            dbt_builtins = {
+                # `is_incremental()` renders as False, always in this case.
+                # TODO: This means we'll never parse the other part of the query,
+                # so we should find a solution to that. Perhaps forcing the file
+                # to be parsed TWICE if it uses this variable.
+                'is_incremental': lambda: False,
+                # Bring through `this` as the first part of the filename
+                'this': this_fname.split('.')[0]
+            }
+            for name in dbt_builtins:
+                # Only apply if it hasn't already been set at this stage.
+                if name not in live_context:
+                    live_context[name] = dbt_builtins[name]
 
+        # Load config macros
+        ctx = self._extract_macros_from_config(config=config, env=env, ctx=live_context)
+        # Load macros from path (if applicable)
+        macros_path = config.get_section((self.templater_selector, self.name, 'load_macros_from_path'))
+        if macros_path:
+            ctx.update(self._extract_macros_from_path(macros_path, env=env, ctx=live_context))
+        live_context.update(ctx)
+
+        # Load the template, passing the global context.
+        template = env.from_string(in_str, globals=live_context)
         violations = []
 
         # Attempt to identify any undeclared variables
@@ -260,12 +319,12 @@ class JinjaTemplateInterface(PythonTemplateInterface):
                             "Undefined jinja template variable: {0!r}".format(tree.name),
                             pos=FilePositionMarker(None, line_no, pos, charpos)
                         )
-
             for val in _crawl_tree(ast, undefined_variables, in_str):
                 violations.append(val)
 
         try:
-            out_str = template.render(**live_context)
+            # NB: Passing no context. Everything is loaded when the template is loaded.
+            out_str = template.render()
             return out_str, violations
         except Exception as err:
             # TODO: Add a url here so people can get more help.